# Conflicts:
#	integrationTests/mock/txProcessorMock.go
#	integrationTests/multiShard/block/executingMiniblocks_test.go
#	integrationTests/multiShard/smartContract/executingSCCalls_test.go
#	process/block/preprocess/transactions.go
#	process/block/shardblock_test.go
#	process/coordinator/process.go
#	process/coordinator/process_test.go
#	process/errors.go
#	process/interface.go
#	process/mock/txProcessorMock.go
#	process/smartContract/export_test.go
#	process/smartContract/process.go
#	process/smartContract/process_test.go
#	process/transaction/process.go

+++ consensus/mock/rounderMock.go
@@ -5,16 +5,16 @@ import (
 )
 
 type RounderMock struct {
-	RoundIndex int32
+	RoundIndex int64
 
-	IndexCalled         func() int32
+	IndexCalled         func() int64
 	TimeDurationCalled  func() time.Duration
 	TimeStampCalled     func() time.Time
 	UpdateRoundCalled   func(time.Time, time.Time)
 	RemainingTimeCalled func(startTime time.Time, maxTime time.Duration) time.Duration
 }
 
-func (rndm *RounderMock) Index() int32 {
+func (rndm *RounderMock) Index() int64 {
 	if rndm.IndexCalled != nil {
 		return rndm.IndexCalled()
 	}
